requirements.txt
flask==3.0.2
werkzeug==3.0.1
jinja2==3.1.3
python-dotenv==1.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80
spacy==3.7.2
itsdangerous==2.1.2
flask-cors==4.0.0
